Adding changes based on the second set of feedback
Added changes based on the set of feedback received.

Changed labels and added more clarity to graphs.

Changed the last question's graph to a scatterplot as it was a histogram.

Changed the first Question to be more specific in what I was searching for.

Updated my summary and conclusion based on my new findings and observations from my changes.

Added a link to my repository.